Fix bug in body-generating function
Previously, when generating the body from text_vector, I had
repeatedly appended new dictionaries to the list of body documents,
where each dictionary would have a new index and the next word in the
vector. Unfortunately, this was both incorrect and not idiomatic, so I
changed it to a functioning list comprehension.

Also added a bit about expected output.